refactor: absorb invite-codes into Settings, delete old pages
Delete 9 files no longer needed:
- AppLayout.tsx (replaced by ChatLayout)
- ChatsLayout.tsx (replaced by ChatLayout + ConversationList)
- ChatsEmptyState.tsx (replaced by ChatLayout empty state)
- TasksPage.tsx (tasks removed from nav)
- ResourcesPage.tsx (replaced by marketplace)
- ContactsPage.tsx (replaced by contacts/ContactList)
- LibraryPage.tsx (already unused)
- InviteCodesPage.tsx (absorbed into SettingsPage)
- Sidebar.tsx (replaced by ConversationList)

Absorb invite codes management into SettingsPage as a new section.
