--- frontend/app/src/components/Sidebar.tsx
@@ -1,477 +1,0 @@
-import { Check, ChevronRight, MoreHorizontal, Plus, Search, Trash2 } from "lucide-react";
-import { useEffect, useMemo, useState } from "react";
-import { Link, useLocation, useParams } from "react-router-dom";
-import type { ThreadSummary } from "../api";
-import MemberAvatar from "./MemberAvatar";
-import { useAppStore } from "../store/app-store";
-import { Skeleton } from "./ui/skeleton";
-
-function requireThreadMemberId(thread: ThreadSummary): string {
-  // @@@thread-member-id-required - thread grouping/routing must use stable member IDs, never display names.
-  if (!thread.member_id) {
-    throw new Error(`Thread ${thread.thread_id} missing member_id`);
-  }
-  return thread.member_id;
-}
-
-function requireSidebarLabel(thread: ThreadSummary): string {
-  if (thread.is_main) {
-    throw new Error(`Main thread ${thread.thread_id} should not render as child thread`);
-  }
-  if (!thread.sidebar_label) {
-    throw new Error(`Thread ${thread.thread_id} missing sidebar_label`);
-  }
-  return thread.sidebar_label;
-}
-
-function memberThreadHref(memberId: string, mainThreadId?: string): string {
-  const encodedMemberId = encodeURIComponent(memberId);
-  // @@@main-thread-direct-route - sidebar switching should reuse the known main
-  // thread route directly; bouncing through /threads/:memberId remounts
-  // NewChatPage and re-runs member bootstrap before landing in ChatPage.
-  return mainThreadId
-    ? `/threads/${encodedMemberId}/${mainThreadId}`
-    : `/threads/${encodedMemberId}`;
-}
-
-function formatRelativeTime(dateStr?: string): string {
-  if (!dateStr) return "";
-  const date = new Date(dateStr);
-  if (isNaN(date.getTime())) return "";
-  const now = new Date();
-  const diffMs = now.getTime() - date.getTime();
-  const diffMinutes = Math.floor(diffMs / 60000);
-  const diffHours = Math.floor(diffMs / 3600000);
-  const diffDays = Math.floor(diffMs / 86400000);
-  if (diffMinutes < 1) return "刚刚";
-  if (diffMinutes < 60) return `${diffMinutes}分钟前`;
-  if (diffHours < 24) return `${diffHours}小时前`;
-  const todayStart = new Date(now.getFullYear(), now.getMonth(), now.getDate());
-  const yesterdayStart = new Date(todayStart.getTime() - 86400000);
-  if (date >= yesterdayStart && date < todayStart) return "昨天";
-  if (diffDays < 7) return `${diffDays}天前`;
-  return `${date.getMonth() + 1}月${date.getDate()}日`;
-}
-
-interface SidebarProps {
-  threads: ThreadSummary[];
-  collapsed?: boolean;
-  loading?: boolean;
-  width?: number;
-  onDeleteThread: (threadId: string) => void;
-  onSearchClick: () => void;
-  onNewChat: () => void;
-}
-
-function ThreadSkeleton() {
-  return (
-    <div className="space-y-0.5">
-      {[...Array(5)].map((_, i) => (
-        <div key={i} className="px-3 py-2.5 rounded-lg" style={{ animationDelay: `calc(var(--duration-instant) * ${i})` }}>
-          <Skeleton className="h-4 w-[70%] mb-1.5" />
-          <Skeleton className="h-3 w-[40%]" />
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function ThreadItem({
-  thread,
-  isActive,
-  label,
-  to,
-  isSelectMode,
-  isSelected,
-  onToggleSelect,
-  confirmDelete,
-  setConfirmDelete,
-  onDeleteThread,
-}: {
-  thread: ThreadSummary;
-  isActive: boolean;
-  label: string;
-  to: string;
-  isSelectMode: boolean;
-  isSelected: boolean;
-  onToggleSelect: (id: string) => void;
-  confirmDelete: string | null;
-  setConfirmDelete: (id: string | null) => void;
-  onDeleteThread: (id: string) => void;
-}) {
-  return (
-    <div className={`group/item flex items-center rounded-lg transition-colors duration-fast ${
-      isSelected ? "bg-primary/10" : isActive ? "bg-background shadow-sm" : "hover:bg-muted"
-    }`}>
-      {/* Left gutter: fixed w-7, holds active indicator OR checkbox — text never moves */}
-      <div className="relative w-7 flex-shrink-0 self-stretch flex items-center justify-center">
-        {/* Active indicator line */}
-        {isActive && !isSelected && (
-          <div className="absolute left-0 top-2 bottom-2 w-0.5 rounded-r-full bg-foreground" />
-        )}
-        {isSelected && (
-          <div className="absolute left-0 top-2 bottom-2 w-0.5 rounded-r-full bg-primary" />
-        )}
-        {/* Checkbox — only visible in select mode */}
-        {isSelectMode && (
-          <button
-            className={`w-4 h-4 rounded border-[1.5px] flex items-center justify-center transition-colors duration-fast ${
-              isSelected ? "bg-primary border-primary" : "border-muted-foreground/40 bg-card"
-            }`}
-            onClick={(e) => { e.stopPropagation(); onToggleSelect(thread.thread_id); }}
-          >
-            {isSelected && <Check className="w-2.5 h-2.5 text-primary-foreground" />}
-          </button>
-        )}
-      </div>
-
-      {/* Text content */}
-      <Link
-        to={isSelectMode ? "#" : to}
-        onClick={(e) => { if (isSelectMode) { e.preventDefault(); onToggleSelect(thread.thread_id); } }}
-        className="flex-1 min-w-0 py-2.5 pr-2"
-      >
-        <div className={`flex items-center gap-1.5 ${isActive ? "text-foreground font-medium" : "text-foreground"}`}>
-          {thread.running && !isSelectMode && (
-            <span className="w-2 h-2 rounded-full bg-success flex-shrink-0 animate-pulse" />
-          )}
-          <span className="text-sm font-medium truncate">{label}</span>
-        </div>
-        <div className="flex items-center gap-1 mt-0.5">
-          <span className="text-xs text-muted-foreground/60 truncate flex-1 min-w-0">
-            {thread.sandbox || "local"}
-          </span>
-          {thread.updated_at && (
-            <span className="text-2xs text-muted-foreground/40 flex-shrink-0">
-              {formatRelativeTime(thread.updated_at)}
-            </span>
-          )}
-        </div>
-      </Link>
-
-      {/* Single-item delete — hidden in select mode */}
-      {!isSelectMode && (
-        <div className={`${confirmDelete === thread.thread_id ? "flex" : "hidden group-hover/item:flex"} items-center gap-0.5 pr-1.5`}>
-          {confirmDelete === thread.thread_id ? (
-            <>
-              <button
-                className="w-6 h-6 rounded flex items-center justify-center text-destructive bg-destructive/10 hover:bg-destructive/20"
-                onClick={(e) => { e.stopPropagation(); setConfirmDelete(null); onDeleteThread(thread.thread_id); }}
-              >
-                <Trash2 className="w-3.5 h-3.5" />
-              </button>
-              <button
-                className="w-6 h-6 rounded flex items-center justify-center text-muted-foreground/60 hover:bg-muted hover:text-foreground text-xs"
-                onClick={(e) => { e.stopPropagation(); setConfirmDelete(null); }}
-              >
-                ✕
-              </button>
-            </>
-          ) : (
-            <button
-              className="w-6 h-6 rounded flex items-center justify-center text-muted-foreground/60 hover:bg-muted hover:text-foreground"
-              onClick={(e) => { e.stopPropagation(); setConfirmDelete(thread.thread_id); }}
-            >
-              <MoreHorizontal className="w-3.5 h-3.5" />
-            </button>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-export default function Sidebar({
-  threads,
-  collapsed = false,
-  loading = false,
-  width = 272,
-  onDeleteThread,
-  onSearchClick,
-  onNewChat,
-}: SidebarProps) {
-  const location = useLocation();
-  const { memberId, threadId } = useParams<{ memberId?: string; threadId?: string }>();
-  const activeMemberId = memberId ? decodeURIComponent(memberId) : null;
-  const activeThreadId = threadId || null;
-  const [confirmDelete, setConfirmDelete] = useState<string | null>(null);
-  const [expandedMembers, setExpandedMembers] = useState<Set<string>>(() => {
-    try {
-      const saved = localStorage.getItem("sidebar-expanded-members");
-      return saved ? new Set(JSON.parse(saved)) : new Set();
-    } catch { return new Set(); }
-  });
-  const [isSelectMode, setIsSelectMode] = useState(false);
-  const [selectedIds, setSelectedIds] = useState<Set<string>>(new Set());
-
-  const onToggleSelect = (threadId: string) => {
-    if (!isSelectMode) setIsSelectMode(true);
-    setSelectedIds(prev => {
-      const next = new Set(prev);
-      if (next.has(threadId)) next.delete(threadId);
-      else next.add(threadId);
-      return next;
-    });
-  };
-
-  const exitSelectMode = () => { setIsSelectMode(false); setSelectedIds(new Set()); };
-
-  const isAllSelected = threads.length > 0 && threads.every(t => selectedIds.has(t.thread_id));
-
-  const handleSelectAll = () => {
-    setSelectedIds(isAllSelected ? new Set() : new Set(threads.map(t => t.thread_id)));
-  };
-
-  const handleBulkDelete = () => {
-    selectedIds.forEach(id => onDeleteThread(id));
-    exitSelectMode();
-  };
-
-  useEffect(() => {
-    if (!isSelectMode) return;
-    const onKey = (e: KeyboardEvent) => { if (e.key === "Escape") exitSelectMode(); };
-    document.addEventListener("keydown", onKey);
-    return () => document.removeEventListener("keydown", onKey);
-  }, [isSelectMode]);
-
-  const memberList = useAppStore(s => s.memberList);
-
-  // Group threads by member, then merge in members with no threads
-  const groups = useMemo(() => {
-    const map = new Map<string, { memberName: string; avatarUrl?: string; threads: ThreadSummary[]; latestAt: number }>();
-
-    for (const thread of threads) {
-      const key = requireThreadMemberId(thread);
-      if (!map.has(key)) map.set(key, { memberName: thread.member_name || "Agent", avatarUrl: thread.avatar_url, threads: [], latestAt: 0 });
-      const g = map.get(key)!;
-      if (!g.memberName && thread.member_name) g.memberName = thread.member_name;
-      if (!g.avatarUrl && thread.avatar_url) g.avatarUrl = thread.avatar_url;
-      const at = thread.updated_at ? new Date(thread.updated_at).getTime() : 0;
-      g.threads.push(thread);
-      g.latestAt = Math.max(g.latestAt, at);
-    }
-
-    // Add members that have no threads yet (e.g. newly created copies)
-    for (const member of memberList) {
-      if (!map.has(member.id)) {
-        map.set(member.id, { memberName: member.name, avatarUrl: member.avatar_url, threads: [], latestAt: 0 });
-      }
-    }
-
-    return [...map.entries()]
-      .map(([memberId, g]) => ({ memberId, ...g }))
-      .sort((a, b) => b.latestAt - a.latestAt)
-      .map(g => ({
-        ...g,
-        threads: [...g.threads].sort((a, b) => {
-          const ta = a.updated_at ? new Date(a.updated_at).getTime() : 0;
-          const tb = b.updated_at ? new Date(b.updated_at).getTime() : 0;
-          return tb - ta;
-        }),
-      }));
-  }, [threads, memberList]);
-
-  const toggleMember = (memberId: string) => {
-    setExpandedMembers(prev => {
-      const next = new Set(prev);
-      if (next.has(memberId)) next.delete(memberId);
-      else next.add(memberId);
-      localStorage.setItem("sidebar-expanded-members", JSON.stringify([...next]));
-      return next;
-    });
-  };
-
-  function isMemberActive(memberId: string, mainThreadId?: string): boolean {
-    if (memberId !== activeMemberId) return false;
-    if (location.pathname === `/threads/${encodeURIComponent(memberId)}/new`) return false;
-    return !activeThreadId || activeThreadId === mainThreadId;
-  }
-
-  // ── Collapsed (narrow) mode ──────────────────────────────────────────────
-  if (collapsed) {
-    return (
-      <div className="w-14 h-full flex flex-col items-center py-3 bg-card border-r border-border animate-slide-in overflow-hidden flex-shrink-0">
-        <button onClick={onNewChat} className="w-9 h-9 rounded-lg flex items-center justify-center mb-1 text-muted-foreground hover:bg-muted hover:text-foreground">
-          <Plus className="w-4 h-4" />
-        </button>
-        <button onClick={onSearchClick} className="w-9 h-9 rounded-lg flex items-center justify-center mb-2 text-muted-foreground hover:bg-muted hover:text-foreground">
-          <Search className="w-4 h-4" />
-        </button>
-
-        <div className="w-8 h-px bg-border mb-2" />
-
-        <div className="flex-1 min-h-0 overflow-y-auto w-full flex flex-col items-center gap-1 px-2 py-1 custom-scrollbar">
-          {groups.map((group) => {
-            const mainThread = group.threads.find((thread) => thread.is_main);
-            const isActive = isMemberActive(group.memberId, mainThread?.thread_id);
-            const isRunning = group.threads.some(t => t.running);
-            return (
-              <div key={group.memberId} className="relative group/item w-full flex justify-center">
-                <Link
-                  to={memberThreadHref(group.memberId, mainThread?.thread_id)}
-                  title={group.memberName}
-                  className={`flex items-center justify-center rounded-xl p-1 transition-colors duration-fast ${
-                    isActive ? "bg-muted" : "hover:bg-muted/70"
-                  }`}
-                >
-                  {isRunning
-                    ? <span className="w-9 h-9 rounded-xl flex items-center justify-center bg-muted"><span className="w-3 h-3 rounded-full border-2 border-muted-foreground border-t-transparent animate-spin" /></span>
-                    : <MemberAvatar name={group.memberName} avatarUrl={group.avatarUrl} type="mycel_agent" size="sm" />}
-                </Link>
-                <div className="absolute left-[52px] top-1/2 -translate-y-1/2 px-2 py-1 bg-foreground text-background text-xs rounded opacity-0 group-hover/item:opacity-100 pointer-events-none transition-opacity duration-fast whitespace-nowrap z-50 max-w-[200px] truncate">
-                  {group.memberName}
-                </div>
-              </div>
-            );
-          })}
-        </div>
-      </div>
-    );
-  }
-
-  // ── Expanded mode ────────────────────────────────────────────────────────
-
-  return (
-    <div className="h-full flex flex-col bg-card border-r border-border animate-slide-in flex-shrink-0" style={{ width }}>
-      {/* Header */}
-      <div className="px-4 pt-3 pb-1 flex items-center justify-between">
-        <span className="text-sm font-semibold text-foreground">消息</span>
-      </div>
-
-      {/* Search */}
-      <div className="px-3 pb-3">
-        <button
-          className="w-full flex items-center gap-2 px-3 py-2 rounded-lg text-sm text-muted-foreground/60 hover:bg-muted hover:text-foreground"
-          onClick={onSearchClick}
-        >
-          <Search className="w-4 h-4" />
-          <span>搜索对话...</span>
-        </button>
-      </div>
-
-      <div className="h-px mx-3 bg-border" />
-
-      {/* Bulk action bar */}
-      {isSelectMode && (
-        <div className="px-3 py-2.5 border-b border-border flex items-center gap-2 flex-shrink-0">
-          <button
-            onClick={handleSelectAll}
-            className="text-xs text-muted-foreground/70 hover:text-foreground transition-colors duration-fast"
-          >
-            {isAllSelected ? "取消全选" : "全选"}
-          </button>
-          <span className="text-xs text-muted-foreground/40">·</span>
-          <span className="text-xs text-muted-foreground flex-1">已选 {selectedIds.size} 条</span>
-          <button
-            onClick={handleBulkDelete}
-            disabled={selectedIds.size === 0}
-            className="flex items-center gap-1 px-2.5 py-1.5 rounded-lg bg-destructive/10 text-destructive hover:bg-destructive/20 disabled:opacity-40 text-xs font-medium transition-colors duration-fast"
-          >
-            <Trash2 className="w-3 h-3" />
-            删除
-          </button>
-          <button
-            onClick={exitSelectMode}
-            className="px-2.5 py-1.5 rounded-lg text-xs text-muted-foreground hover:bg-muted transition-colors duration-fast"
-          >
-            取消
-          </button>
-        </div>
-      )}
-
-      {/* Thread list */}
-      <div className="flex-1 min-h-0 px-3 pt-3 flex flex-col">
-        <div className="flex items-center justify-between px-2 mb-2 flex-shrink-0">
-          <span className="text-xs font-medium tracking-wider uppercase text-muted-foreground/60">对话</span>
-          <div className="flex items-center gap-1.5">
-            <span className="text-xs text-muted-foreground/40">{threads.length}</span>
-            {!isSelectMode && (
-              <button
-                onClick={() => setIsSelectMode(true)}
-                className="text-xs text-muted-foreground/50 hover:text-foreground transition-colors duration-fast px-1"
-              >
-                管理
-              </button>
-            )}
-          </div>
-        </div>
-
-        <div className="flex-1 min-h-0 overflow-y-auto space-y-0.5 custom-scrollbar">
-          {loading ? (
-            <ThreadSkeleton />
-          ) : (
-            groups.map((group) => {
-              const isExpanded = expandedMembers.has(group.memberId);
-              const urlId = encodeURIComponent(group.memberId);
-              const mainThread = group.threads.find((thread) => thread.is_main);
-              const memberHref = memberThreadHref(group.memberId, mainThread?.thread_id);
-              const memberIsActive = isMemberActive(group.memberId, mainThread?.thread_id);
-              const childThreads = group.threads.filter((thread) => !thread.is_main);
-              return (
-                <div key={group.memberId} className="mb-1">
-                  <div className={`flex items-center gap-1 px-2 py-1.5 rounded-xl transition-colors duration-fast ${
-                    memberIsActive
-                      ? "bg-muted"
-                      : "hover:bg-muted/70"
-                  }`}>
-                    <button
-                      onClick={() => toggleMember(group.memberId)}
-                      className={`w-5 h-5 flex items-center justify-center rounded transition-colors duration-fast ${
-                        memberIsActive ? "hover:bg-background/80" : "hover:bg-background/60"
-                      }`}
-                      aria-label={isExpanded ? "收起分支对话" : "展开分支对话"}
-                    >
-                      <ChevronRight className={`w-3.5 h-3.5 transition-transform duration-fast flex-shrink-0 ${
-                        memberIsActive ? "text-foreground/70" : "text-muted-foreground/50"
-                      } ${isExpanded ? "rotate-90" : ""}`} />
-                    </button>
-                    <Link
-                      to={memberHref}
-                      className="flex items-center gap-1.5 min-w-0 flex-1"
-                    >
-                      <MemberAvatar name={group.memberName} avatarUrl={group.avatarUrl} type="mycel_agent" size="xs" />
-                      <span className={`text-xs flex-1 truncate ${
-                        memberIsActive ? "font-semibold text-foreground" : "font-medium text-foreground"
-                      }`}>
-                        {group.memberName}
-                      </span>
-                    </Link>
-                  </div>
-                  {isExpanded && (
-                    <>
-                      <div className="mt-0.5 ml-3 space-y-0.5">
-                        {childThreads.map((thread) => (
-                          <ThreadItem
-                            key={thread.thread_id}
-                            thread={thread}
-                            isActive={activeThreadId === thread.thread_id}
-                            label={requireSidebarLabel(thread)}
-                            to={`/threads/${urlId}/${thread.thread_id}`}
-                            isSelectMode={isSelectMode}
-                            isSelected={selectedIds.has(thread.thread_id)}
-                            onToggleSelect={onToggleSelect}
-                            confirmDelete={confirmDelete}
-                            setConfirmDelete={setConfirmDelete}
-                            onDeleteThread={onDeleteThread}
-                          />
-                        ))}
-                      </div>
-                      <div className="px-3">
-                        <Link
-                          to={`/threads/${urlId}/new`}
-                          className="block py-2 text-center text-xs text-muted-foreground/50 hover:text-muted-foreground transition-colors duration-fast"
-                        >
-                          + 发起新对话
-                        </Link>
-                      </div>
-                    </>
-                  )}
-                </div>
-              );
-            })
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
