--- frontend/app/src/pages/ChatsEmptyState.tsx
@@ -1,7 +1,0 @@
-export default function ChatsEmptyState() {
-  return (
-    <div className="h-full flex items-center justify-center">
-      <p className="text-sm text-muted-foreground">选择一个对话或发起新会话</p>
-    </div>
-  );
-}
